PROTON-1774 Remove sasl processing when complete
Once complete remove any impact of the SaslTransportWrapper from the
transport chain by selecting back to the next TransportWrappers for
input and output once SASL is complete and the pending buffers are
drained.  Move the buffers for sasl input and output into the wrapper
which can be GC'd once removed from the chain.