PROTON-1941 Add ensureRemaining to the WritableBuffer interface
Provide an access point for the codec to request that the WritableBuffer
be sized to accept a given number of bytes usually in relation to larger
incoming write of a complex type.  Add default implementation  to preseve
backwards compatability and an implementation in the ByteBuffer wrapper
that throws early if nont enough bytes exist to handle the announced
pending writes.