PROTON-1948 Refactor FrameWriter with more performant buffer
Refactor the FrameWriter to use a more performant ReadableBuffer wrapper
around a byte array that auto grows as an ecode requires instead of
reallocating a ByteBuffer and copying past contents and then performing
a new encode any time space is exhausted.  Removing the layer of buffer
abstractions leads to an increase in overall performance under normal
operations as well.